flatten: nokhte_blur hierarchy
