+++ packages/client/lib/app/core/widgets/nokhte_blur/mobx/nokhte_blur_store.dart
@@ -2,7 +2,7 @@
 import 'package:mobx/mobx.dart';
 import 'package:nokhte/app/core/mobx/mobx.dart';
 import 'package:nokhte/app/core/types/types.dart';
-import 'package:nokhte/app/core/widgets/nokhte_blur/stack/stack.dart';
+import 'package:nokhte/app/core/widgets/widgets.dart';
 import 'package:simple_animations/simple_animations.dart';
 part 'nokhte_blur_store.g.dart';
