--- packages/client/lib/app/core/widgets/nokhte_blur/stack/presentation/presentation.dart
@@ -1,3 +1,0 @@
-export 'mobx/nokhte_blur_store.dart';
-export 'widget/nokhte_blur.dart';
-export 'movies/blur_movie.dart';
