--- packages/client/lib/app/core/widgets/nokhte_blur/stack/presentation/widget/nokhte_blur.dart
@@ -1,30 +1,0 @@
-import 'package:blur/blur.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:nokhte/app/core/widgets/widgets.dart';
-import 'package:simple_animations/simple_animations.dart';
-
-class NokhteBlur extends StatelessWidget {
-  final NokhteBlurStore store;
-  const NokhteBlur({
-    super.key,
-    required this.store,
-  });
-
-  @override
-  Widget build(BuildContext context) => Observer(
-        builder: (context) => CustomAnimationBuilder(
-          tween: store.movie,
-          control: store.control,
-          duration: store.movie.duration,
-          onCompleted: () => store.onCompleted(),
-          builder: (context, value, child) {
-            return Blur(
-              colorOpacity: 0.0,
-              blur: value.get('blur'),
-              child: Container(),
-            );
-          },
-        ),
-      );
-}
